+++ src/app/app.component.html
@@ -16,13 +16,9 @@
             <li nz-menu-item><a class="routing_link" routerLink="/user-list">list of users</a></li>
             <li nz-menu-item><a class="routing_link" routerLink="/search-user">Search by full name</a></li>
             <li nz-menu-item><a class="routing_link" routerLink="/pro-sub">Project Submission</a></li>
-            <li nz-menu-item><a class="routing_link" routerLink="/pro-sub-2">Project Submission 2</a></li>
             <li nz-menu-item><a class="routing_link" routerLink="/per-mana">Person Management</a></li>
-            <li nz-menu-item><a class="routing_link" routerLink="/per-mana-2">Person Management 2</a></li>
             <li nz-menu-item><a class="routing_link" routerLink="/pro-mana">Project Management</a></li>
-            <li nz-menu-item><a class="routing_link" routerLink="/pro-mana-2">Project Management 2</a></li>
             <li nz-menu-item><a class="routing_link" routerLink="/pro-mem">Project Members</a></li>
-            <li nz-menu-item><a class="routing_link" routerLink="/pro-mem-2">Project Members 2</a></li>
             <li nz-menu-item><a class="routing_link" routerLink="/company-management">Company Management</a></li>
             <li nz-menu-item><a class="routing_link" routerLink="/company-managementv2">Company ManagementV2</a></li>
             <li nz-menu-item><a class="routing_link" routerLink="/user-group">User Group</a></li>
