--- src/app/project-members-v2/project-members-v2.component.html
@@ -1,130 +1,0 @@
-<div class="container">
-  <div class="jumbotron">
-    <h1>Project Members</h1>
-    <form nz-form [nzLayout]="'inline'" [formGroup]="validateForm" (ngSubmit)="onSubmit()">
-      <div nz-form-item>
-        <div nz-form-control [nzValidateStatus]="validateForm.controls.project_id">
-          <nz-select formControlName="project_id" [(ngModel)]="model.project_id" [nzShowSearch]="true" [nzPlaceHolder]="'Select a Project ID'" style="width:200px" nzAllowClear>
-            <nz-option *ngFor="let project of projects | async" [nzLabel]="project._id" [nzValue]="project._id">
-            </nz-option>
-          </nz-select>
-          <div nz-form-explain *ngIf="validateForm.controls.project_id.dirty&&validateForm.controls.project_id.hasError('required')">Please input your Project ID!</div>
-        </div>
-      </div>
-      <div nz-form-item>
-        <div nz-form-control [nzValidateStatus]="validateForm.controls.person_id">
-          <nz-select formControlName="person_id" [(ngModel)]="model.person_id" [nzShowSearch]="true" [nzPlaceHolder]="'Select a Person ID'" style="width:200px" nzAllowClear>
-            <nz-option *ngFor="let person of persons | async" [nzLabel]="person._id" [nzValue]="person._id">
-            </nz-option>
-          </nz-select>
-          <div nz-form-explain *ngIf="validateForm.controls.person_id.dirty&&validateForm.controls.person_id.hasError('required')">Please input your Person ID!</div>
-        </div>
-      </div>
-      <button nz-button [nzType]="'primary'" [nzSize]="'large'" [disabled]="!validateForm.valid">Submit</button>
-    </form>
-  </div>
-
-  <div *ngIf="(result)?.length > 0" class="row">
-    <div class="col-xs-12 col-12 col-md-6">
-      <pagination class="pagination-sm" [boundaryLinks]="true" [rotate]="false" [totalItems]="totalItems" [(ngModel)]="currentPage"
-        [itemsPerPage]="itemsPerPage" [maxSize]="maxSize" (pageChanged)="pageChanged($event)" (numPages)="numPages = $event">
-      </pagination>
-    </div>
-  </div>
-
-  <pre *ngIf="(result)?.length > 0" class="card card-block card-header">Page: {{currentPage}} / {{numPages}}</pre>
-
-  <nz-table *ngIf="(result)?.length>0" #nzTable [nzAjaxData]="projectMembers" [nzPageSize]="10" [nzShowSizeChanger]="true">
-    <thead nz-thead>
-      <tr>
-        <th nz-th>
-          <span>#</span>
-        </th>
-        <th nz-th>
-          <span>isEnabled</span>
-        </th>
-        <th nz-th>
-          <span>Name</span>
-        </th>
-        <th nz-th>
-          <span>Project</span>
-        </th>
-        <th nz-th>
-          <span>Company</span>
-        </th>
-        <th nz-th>
-          <span>Country</span>
-        </th>
-        <th nz-th>
-          <span>Created Date</span>
-        </th>
-        <th nz-th>
-          <span>Updated Date</span>
-        </th>
-        <th nz-th>
-          <span>Actions</span>
-        </th>
-      </tr>
-    </thead>
-    <tbody nz-tbody>
-      <tr nz-tbody-tr *ngFor="let projectMember of result; let i=index">
-        <td nz-td>{{projectMember._id}}</td>
-        <td nz-td>
-          <nz-switch [ngModel]="projectMember.isEnabled" (click)="onEnable(projectMember)">
-          </nz-switch>
-        </td>
-        <td nz-td>{{projectMember.person_id}}</td>
-        <td nz-td>{{projectMember.project_id}}</td>
-        <td nz-td>{{projectMember.company_id}}</td>
-        <td nz-td>{{projectMember.isEnabled}}</td>
-        <td nz-td>{{projectMember.created_date}}</td>
-        <td nz-td>{{projectMember.modified_date}}</td>
-        <td nz-td>
-          <button nz-button [nzType]="'primary'" (click)="edit(projectMember)">
-            <i class="anticon anticon-edit"></i>
-          </button>
-          <button nz-button [nzType]="'danger'" (click)="onDelete(projectMember)">
-            <i class="anticon anticon-delete"></i>
-          </button>
-        </td>
-      </tr>
-    </tbody>
-  </nz-table>
-  <div *ngIf="(result)?.length==0">
-    <p>No record found.</p>
-  </div>
-</div>
-
-
-<nz-modal [nzVisible]="editProjectMemberModal" [nzTitle]="editTitle" [nzContent]="editContent" [nzFooter]="editFooter" (nzOnCancel)="closeModal('editProjectMemberModal')">
-  <ng-template #editTitle>
-    Edit Project
-  </ng-template>
-  <ng-template #editContent>
-
-    <form nz-form [formGroup]="validateEditForm">
-
-      <div nz-form-item>
-        <div nz-form-label>
-          <label>Project ID</label>
-        </div>
-        <div nz-form-control [nzValidateStatus]="validateEditForm.controls.project_id">
-          <nz-select formControlName="project_id" [(ngModel)]="editProjectMember.project_id" nzAllowClear>
-            <nz-option *ngFor="let project of projects | async" [nzLabel]="project._id" [nzValue]="project._id">
-            </nz-option>
-          </nz-select>
-          <div nz-form-explain *ngIf="validateEditForm.controls.project_id.dirty&&validateEditForm.controls.project_id.hasError('required')">Please input your Project IDß</div>
-        </div>
-      </div>
-
-    </form>
-  </ng-template>
-  <ng-template #editFooter>
-    <button nz-button [nzType]="'default'" [nzSize]="'large'" (click)="closeModal('editProjectMemberModal')">
-      Cancel
-    </button>
-    <button nz-button [nzType]="'primary'" [nzSize]="'large'" [disabled]="!validateEditForm.valid" (click)="onEdit()">
-      Save
-    </button>
-  </ng-template>
-</nz-modal>
